# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Inventory Management API
  version: 1.0.0
  description: |
    Backend for a multi-tenant inventory platform for SMEs, consumed by an
    Android (Kotlin/Compose) client.
  license:
    name: Proprietary
    # SPDX identifier. OpenAPI 3.1 treats this as optional, but swagger-parser
    # reports its absence, and the alternative — teaching ApiContractTest to
    # ignore parser messages — would also silence the dangling-$ref detection
    # that makes the check worth having. LicenseRef-* is the SPDX form for a
    # licence that is not on the SPDX list. Mutually exclusive with license.url,
    # which is why there is no url here.
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /inventory:
    get:
      tags: [Inventory]
      summary: Stock status across products
      description: Backs the main inventory screen and the low-stock badge.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/LocationIdQuery'
        - name: stockState
          in: query
          schema: { type: string, enum: [ok, reorder, out_of_stock, unknown] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/StockStatus' }
  /inventory/movements:
    get:
      tags: [Inventory]
      summary: Read the stock ledger
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: productId
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/LocationIdQuery'
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/MovementType' }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/StockMovement' }
  /inventory/adjustments:
    post:
      tags: [Inventory]
      summary: Correct stock (damage, theft, recount)
      description: |
        Appends an `adjustment` movement. There is deliberately no endpoint to
        set an absolute quantity or to edit a past movement — corrections are
        always new signed rows, so the audit trail cannot be rewritten.

        The `Idempotency-Key` header makes a retry return the original movement
        rather than posting a second one. Essential for an adjustment captured
        offline and replayed on reconnect: the ledger is append-only, so a
        double-post is repaired only by a compensating row somebody has to
        notice is needed.
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdjustmentRequest' }
      responses:
        '201':
          description: Posted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockMovement' }
        '200':
          description: Idempotent replay of an already-posted adjustment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockMovement' }
        '409': { $ref: '#/components/responses/InsufficientStock' }
  /inventory/transfers:
    post:
      tags: [Inventory]
      summary: Move stock between locations
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, fromLocationId, toLocationId, quantity]
              properties:
                productId: { type: string, format: uuid }
                fromLocationId: { type: string, format: uuid }
                toLocationId: { type: string, format: uuid }
                quantity: { type: number, exclusiveMinimum: 0 }
                reason: { type: string }
      responses:
        '201':
          description: Transferred (two paired movements)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StockMovement' }
        '409': { $ref: '#/components/responses/InsufficientStock' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token, 15 minute TTL. Claims: `sub` (user id), `tid` (tenant id),
        `role`, `jti`. The server trusts `tid` from the signed token only.
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Client-generated UUID; a replay returns the original result
      schema: { type: string, format: uuid }
    LocationIdQuery:
      name: locationId
      in: query
      description: Defaults to the tenant's default location
      schema: { type: string, format: uuid }
...
  responses:
...
    InsufficientStock:
      description: The operation would drive stock negative
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/InsufficientStockProblem' }
  schemas:
...
    InsufficientStockProblem:
      description: |
        A 409 from any endpoint that moves stock. Carries the numbers a cashier
        needs to act on: which product, how much was asked for, how much there
        actually is.
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [productId, requested, available]
          properties:
            productId: { type: string, format: uuid }
            requested:
              type: number
              description: Absolute quantity requested, not the signed delta
            available:
              type: number
              description: Quantity on hand at the time of the refusal
    Problem:
      type: object
      description: RFC 9457 problem detail
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ['https://api.example.com/problems/insufficient-stock'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    Page:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { type: object }
        nextCursor:
          type: string
          nullable: true
          description: Null when the last page has been reached
...
    MovementType:
      type: string
      enum: [purchase_receipt, sale, sale_return, purchase_return, adjustment,
             stocktake, transfer_in, transfer_out, opening_balance]
...
    StockStatus:
      type: object
      required: [productId, sku, name, locationId, locationName, quantityOnHand, quantityReserved, quantityAvailable, stockState, updatedAt]
      properties:
        productId: { type: string, format: uuid }
        sku: { type: string }
        name: { type: string }
        locationId: { type: string, format: uuid }
        locationName: { type: string }
        quantityOnHand: { type: number }
        quantityReserved: { type: number }
        quantityAvailable: { type: number }
        reorderPoint: { type: number, nullable: true }
        stockState: { type: string, enum: [ok, reorder, out_of_stock, unknown] }
        daysOfCover: { type: number, nullable: true }
        updatedAt: { type: string, format: date-time }
    StockMovement:
      type: object
      required: [id, productId, productName, locationId, movementType, quantityDelta, occurredAt]
      properties:
        id: { type: integer, format: int64 }
        productId: { type: string, format: uuid }
        productName: { type: string }
        locationId: { type: string, format: uuid }
        movementType: { $ref: '#/components/schemas/MovementType' }
        quantityDelta: { type: number, description: Signed; negative is stock out }
        balanceAfter: { type: number, nullable: true }
        unitCost: { type: number, nullable: true }
        referenceType: { type: string, nullable: true }
        referenceId: { type: string, format: uuid, nullable: true }
        reason: { type: string, nullable: true }
        occurredAt: { type: string, format: date-time }
        createdBy: { type: string, format: uuid, nullable: true }
        createdByName: { type: string, nullable: true }
    AdjustmentRequest:
      type: object
      required: [productId, quantityDelta, reason]
      properties:
        productId: { type: string, format: uuid }
        locationId: { type: string, format: uuid }
        quantityDelta:
          type: number
          description: Signed and non-zero. Use -2 for two damaged units.
        reason: { type: string, minLength: 1 }
        unitCost: { type: number, minimum: 0 }
        occurredAt: { type: string, format: date-time }
...